[SPARK-46504][PS][TESTS] Factor slow tests out of `IndexesTests`
### What changes were proposed in this pull request?
Factor slow tests out of `IndexesTests`

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44485 from zhengruifeng/ps_test_idx_base.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
