+++ dev/sparktestsupport/modules.py
@@ -795,6 +795,12 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.series",
         # unittests
         "pyspark.pandas.tests.indexes.test_base",
+        "pyspark.pandas.tests.indexes.test_asof",
+        "pyspark.pandas.tests.indexes.test_astype",
+        "pyspark.pandas.tests.indexes.test_delete",
+        "pyspark.pandas.tests.indexes.test_diff",
+        "pyspark.pandas.tests.indexes.test_insert",
+        "pyspark.pandas.tests.indexes.test_map",
         "pyspark.pandas.tests.indexes.test_base_slow",
         "pyspark.pandas.tests.indexes.test_datetime",
         "pyspark.pandas.tests.indexes.test_datetime_at",
@@ -1071,6 +1077,12 @@ pyspark_pandas_connect_part0 = Module(
         "pyspark.pandas.tests.connect.test_parity_typedef",
         "pyspark.pandas.tests.connect.test_parity_utils",
         "pyspark.pandas.tests.connect.indexes.test_parity_base",
+        "pyspark.pandas.tests.connect.indexes.test_parity_asof",
+        "pyspark.pandas.tests.connect.indexes.test_parity_astype",
+        "pyspark.pandas.tests.connect.indexes.test_parity_delete",
+        "pyspark.pandas.tests.connect.indexes.test_parity_diff",
+        "pyspark.pandas.tests.connect.indexes.test_parity_insert",
+        "pyspark.pandas.tests.connect.indexes.test_parity_map",
         "pyspark.pandas.tests.connect.indexes.test_parity_align",
         "pyspark.pandas.tests.connect.indexes.test_parity_indexing",
         "pyspark.pandas.tests.connect.indexes.test_parity_reindex",
